Fai 1065 add a job

+++ src/SFA.DAS.CandidateAccount.Domain/Application/WorkHistoryType.cs
@@ -1,6 +1,6 @@
 ﻿namespace SFA.DAS.CandidateAccount.Domain.Application
 {
-    public enum WorkHistoryType : short
+    public enum WorkHistoryType : byte
     {
         Job = 0,
         WorkExperience = 1
